requirements.txt
PyGithub==2.1.1
openai==1.3.0
python-dotenv==1.0.0
